Makefile: add setup of U-Boot env into firmware-image target
Instead of relying on what the U-Boot 'default' built-in environment is
we will pre-populate it when we create the firmware-image.

This also makes it easier for users to understand where they can alter
the environment in their firmware images.

Note that the newport.env script is not setting a 'blank' environment but
instead altering what is built into U-boot per the U-boot config. If you
need to ensure a variable is not set you need to set it to nothing.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>