Makefile: add ubuntu-image support
With the switch to using U-Boot 'Distro Config' we need to use a
bootscript for Ubuntu. Additionally we will choose to use a compressed
kernel in a FIT image placed on the FATFS now that we have additional
space there. This provides the added benefit of being able to use
root filesystems that are not supported directly by U-Boot such as F2FS.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>