openwrt.scr/ubuntu.scr: update scripts for using scsi vs sata
U-Boot 2019.10 for OcteonTX uses the scsi suybsystem for AHCI over PCI
and not the sata subsystem thus we now use the 'scsi' command and
interface type instead of 'sata'.

Make the bootscripts agnostic to sata/scsi as the only devtype that
needs special handling is the mmc device with its mmcblk<dev>p<part>
naming scheme. The other block storage devices supported use sd{a,b,c}
format.

Note that using PARTUUID would eliminate having to come up with the root
device by path and also be able to elimiante usb/scsi device confusion
but that requires moving from MBR to GPT which we do not use at this time.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>