bootscripts: disable USB autosuspend
The CN81XX errata 29646 states:

Many operating systems support auto-suspend for USB devices, where a device
that has been idle for some period of time is placed into a low-power state.

When USB auto-suspend mode is enabled, it is possible for devices to not be
correctly detected as connected after boot-up or power-on cycles

As a work-around we disable auto-suspend mode.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>