Makefile: fix generation of firmware-newport.bin JTAG recovery image
JTAG recovery images such as firmware-newport.bin are for recovering a
boot image that won't let you get to a working U-Boot prompt with eMMC
and network support.

Therefore the intention is to overwrite everything that the boot
firmware requires which includes the partition table (as it uses the
first FAT32 partition for various boot firmware components) and U-Boot
env.

Fix the generation of firmware-newport.bin:
- do not erase the entire hardware partition (use erase_part vs erase_all)
- fix the specification of start-end - for eMMC these are in units
  of 512byte blocks, not bytes

Signed-off-by: Tim Harvey <tharvey@gateworks.com>