using uncompressed kernel Image instead of kernel.itb FIT image
To reduce complexity switch over to including the uncompressed
kernel Image in the root filesystem and using booti instead of using
a FIT image to compress the kernel. The extra complexity is not worth
the minor disk savings and there is no savings in boot time based on
size/compression.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>